remove move_tensor_sharing_pass, test=develop
