+++ paddle/fluid/framework/details/share_tensor_buffer_op_handle.h
@@ -17,7 +17,9 @@
 #include <unordered_map>
 #include <utility>
 #include <vector>
+#include "paddle/fluid/framework/details/computation_op_handle.h"
 #include "paddle/fluid/framework/details/op_handle_base.h"
+#include "paddle/fluid/framework/details/share_tensor_buffer_functor.h"
 
 namespace paddle {
 namespace framework {
@@ -31,38 +33,6 @@ class MemOptVarInfo;
 }  // namespace ir
 
 namespace details {
-
-class ShareTensorBufferFunctor {
- public:
-  ShareTensorBufferFunctor(Scope *scope, size_t scope_idx,
-                           const std::string &op_type,
-                           const std::vector<ir::MemOptVarInfo *> &in_var_infos,
-                           const std::vector<std::string> &out_var_names);
-
-  void Add(ir::MemOptVarInfo *in_var_info, const std::string &out_var_name);
-
-  void operator()(Scope *exec_scope);
-
-  std::unordered_map<std::string, std::string> ReusedVars() const;
-
-  size_t GetScopeIdx() const { return scope_idx_; }
-
-  Scope *GetScope() { return scope_; }
-
- private:
-  void CallOnce();
-
- private:
-  Scope *scope_;
-  Scope *exec_scope_{nullptr};
-
-  size_t scope_idx_;
-  std::string op_type_;
-  std::vector<ir::MemOptVarInfo *> in_var_infos_;
-  std::vector<std::string> out_var_names_;
-
-  std::vector<std::pair<const Variable *, Variable *>> in_out_vars_;
-};
 
 class ShareTensorBufferOpHandle : public OpHandleBase {
  public:
@@ -96,6 +66,9 @@ class ShareTensorBufferOpHandle : public OpHandleBase {
  private:
   ShareTensorBufferFunctor functor_;
 };
+
+ComputationOpHandle *GetUniquePendingComputationOpHandle(
+    ShareTensorBufferOpHandle *share_tensor_op);
 
 }  // namespace details
 }  // namespace framework
