+++ paddle/fluid/framework/parallel_executor.cc
@@ -242,11 +242,6 @@ ir::Graph *ParallelExecutorPrivate::ApplyMemoryOptimizePass(ir::Graph *graph) {
   VLOG(10) << "ReferenceCountPass Applied";
 
   if (build_strategy_.enable_inplace_) {
-    VLOG(10) << "Before applying buffer_shared_inplace_pass: op num: "
-             << ir::FilterByNodeWrapper<details::OpHandleBase>(*graph).size()
-             << ", var num: "
-             << ir::FilterByNodeWrapper<details::VarHandleBase>(*graph).size()
-             << ", node num: " << graph->Nodes().size();
     auto inplace_pass =
         ir::PassRegistry::Instance().Get("buffer_shared_inplace_pass");
     inplace_pass->SetNotOwned(ir::kMemOptVarInfoMapList, &mem_opt_var_infos_);
@@ -255,11 +250,6 @@ ir::Graph *ParallelExecutorPrivate::ApplyMemoryOptimizePass(ir::Graph *graph) {
     VLOG(10) << "Start to apply buffer_shared_inplace_pass";
     graph = inplace_pass->Apply(graph);
     VLOG(10) << "buffer_shared_inplace_pass Applied";
-    VLOG(10) << "After applying buffer_shared_inplace_pass: op num: "
-             << ir::FilterByNodeWrapper<details::OpHandleBase>(*graph).size()
-             << ", var num: "
-             << ir::FilterByNodeWrapper<details::VarHandleBase>(*graph).size()
-             << ", node num: " << graph->Nodes().size();
   }
 
   if (build_strategy_.memory_optimize_) {
@@ -274,25 +264,6 @@ ir::Graph *ParallelExecutorPrivate::ApplyMemoryOptimizePass(ir::Graph *graph) {
     VLOG(10) << "Start to apply buffer_shared_cross_op_memory_reuse_pass";
     graph = cross_op_memory_reuse_pass->Apply(graph);
     VLOG(10) << "buffer_shared_cross_op_memory_reuse_pass Applied";
-  }
-
-  if (build_strategy_.move_tensor_sharing_to_compute_op_) {
-    VLOG(10)
-        << "Before applying move_tensor_sharing_to_compute_op_pass: op num: "
-        << ir::FilterByNodeWrapper<details::OpHandleBase>(*graph).size()
-        << ", var num: "
-        << ir::FilterByNodeWrapper<details::VarHandleBase>(*graph).size()
-        << ", node num: " << graph->Nodes().size();
-    auto move_tensor_sharing_to_compute_op_pass =
-        ir::PassRegistry::Instance().Get(
-            "move_tensor_sharing_to_compute_op_pass");
-    graph = move_tensor_sharing_to_compute_op_pass->Apply(graph);
-    VLOG(10)
-        << "After applying move_tensor_sharing_to_compute_op_pass: op num: "
-        << ir::FilterByNodeWrapper<details::OpHandleBase>(*graph).size()
-        << ", var num: "
-        << ir::FilterByNodeWrapper<details::VarHandleBase>(*graph).size()
-        << ", node num: " << graph->Nodes().size();
   }
 
   if (GetEagerDeletionThreshold() < 0) {
@@ -822,4 +793,3 @@ USE_PASS(reference_count_pass);
 USE_PASS(eager_deletion_pass);
 USE_PASS(buffer_shared_inplace_pass);
 USE_PASS(buffer_shared_cross_op_memory_reuse_pass);
-USE_PASS(move_tensor_sharing_to_compute_op_pass);
